Tratando excecoes - Exception Handling

+++ src/main/java/com/robertocosta/crudcliente/service/ClientService.java
@@ -1,6 +1,7 @@
 package com.robertocosta.crudcliente.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -9,52 +10,70 @@ import org.springframework.transaction.annotation.Transactional;
 import com.robertocosta.crudcliente.dto.ClientDTO;
 import com.robertocosta.crudcliente.entities.Client;
 import com.robertocosta.crudcliente.repositories.ClientRepository;
+import com.robertocosta.crudcliente.service.exceptions.DatabaseException;
+import com.robertocosta.crudcliente.service.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ClientService {
-	
+
 	@Autowired
 	private ClientRepository repository;
-	
+
 	@Transactional(readOnly = true)
-	public Page<ClientDTO> findAll(Pageable pageable){
+	public Page<ClientDTO> findAll(Pageable pageable) {
 		return repository.findAll(pageable).map(x -> new ClientDTO(x));
 	}
-	
+
 	@Transactional(readOnly = true)
 	public ClientDTO findById(Long id) {
-		return new ClientDTO(repository.findById(id).get());
+		Client client = repository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado!"));
+		return new ClientDTO(client);
 	}
-	
+
 	@Transactional
 	public ClientDTO insert(ClientDTO dto) {
-		Client entity = new Client();		
-		copyDtoToEntity(dto, entity);		
-		entity  = repository.save(entity);
-		return new ClientDTO(entity);
+		try {
+			Client entity = new Client();
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ClientDTO(entity);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("CPF já cadastrado.");
+		}
 	}
-	
+
 	@Transactional
 	public ClientDTO update(Long id, ClientDTO dto) {
-		Client entity = repository.getReferenceById(id);;
-		copyDtoToEntity(dto, entity);		
-		entity  = repository.save(entity);
-		return new ClientDTO(entity);
+		try {
+			Client entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ClientDTO(entity);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado!");
+		} 
+//		catch (DataIntegrityViolationException e) {
+//			throw new DatabaseException("CPF já cadastrado.");
+//		}
 	}
-		
+
 	@Transactional
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado!");
+		}
+			repository.deleteById(id);
 	}
-	
+
 	private void copyDtoToEntity(ClientDTO dto, Client entity) {
 		entity.setName(dto.getName());
 		entity.setCpf(dto.getCpf());
 		entity.setIncome(dto.getIncome());
 		entity.setBirthDate(dto.getBirthDate());
-		entity.setChildren(dto.getChildren());		
+		entity.setChildren(dto.getChildren());
 	}
-
-	
 
 }
